Don't decode message when "Modicon Format" is clicked
Modicon Format is only applicable for Floats and LongInts,
and since we now also support 16-bit integers it's doesn't
make sense to automatically decode the message when this
check box is clicked. Also, it could be kinda confusing for the
user and not really intuitive.

Update the tool tip message to reflect the change.
For some reason VS2015 moves the text out from the resource file
and into the designer file when I update this tool tip text. Not sure why...