Add option to open a mdbus.log file
Instead of having to copy commands one by one from the mdbus.log file,
we now have the option to import a complete file, either by drag-n-drop or
by open file dialog. All the commands will be listed in a listbox with the date
(of the file) and time stamp, and start of the Modbus message.
Clicking on a command will show it in the old text box. Double-clicking a
command will show it in the text box and decode it right away.

Updated version to 1.1.8

* Updated version history file
* changed copyright to 2018
* updated TODO file