Add example LOG files from mdbus.exe
Add log files created from mdbus.exe that can be used to test opening a file
and decode various messages with real data. This includes float- and integer values,
from both Holding and Input registers. It also contains various examples of Illegal Address
error messages when asking for data in registers that doesn't exist.

Mdbus1.log contains some messages with bad checksum (B)
These are actually complete messages that are split up in two
when received by Mdbus. We are not sure why this happens but it might be some timing
issue on the COM port in Windows or timing settings in mdbus.exe config.